Move PG/SF simple schema changes to generic test
BQ would need to implement GetTableSchema interface
