+++ flow/e2e/postgres/peer_flow_pg_test.go
@@ -17,7 +17,6 @@ import (
 	"github.com/PeerDB-io/peer-flow/e2e"
 	"github.com/PeerDB-io/peer-flow/generated/protos"
 	"github.com/PeerDB-io/peer-flow/model"
-	"github.com/PeerDB-io/peer-flow/model/qvalue"
 	"github.com/PeerDB-io/peer-flow/shared"
 	peerflow "github.com/PeerDB-io/peer-flow/workflows"
 )
@@ -49,32 +48,6 @@ func (s PeerFlowE2ETestSuitePG) checkPeerdbColumns(dstSchemaQualified string, ro
 	}
 
 	return nil
-}
-
-func (s PeerFlowE2ETestSuitePG) WaitForSchema(
-	env e2e.WorkflowRun,
-	reason string,
-	srcTableName string,
-	dstTableName string,
-	cols string,
-	expectedSchema *protos.TableSchema,
-) {
-	s.t.Helper()
-	e2e.EnvWaitFor(s.t, env, 3*time.Minute, reason, func() bool {
-		s.t.Helper()
-		output, err := s.conn.GetTableSchema(context.Background(), &protos.GetTableSchemaBatchInput{
-			TableIdentifiers: []string{dstTableName},
-		})
-		if err != nil {
-			return false
-		}
-		tableSchema := output.TableNameSchemaMapping[dstTableName]
-		if !e2e.CompareTableSchemas(expectedSchema, tableSchema) {
-			s.t.Log("schemas unequal", expectedSchema, tableSchema)
-			return false
-		}
-		return s.comparePGTables(srcTableName, dstTableName, cols) == nil
-	})
 }
 
 func (s PeerFlowE2ETestSuitePG) Test_Geospatial_PG() {
@@ -217,188 +190,6 @@ func (s PeerFlowE2ETestSuitePG) Test_Enums_PG() {
 	s.t.Log("Inserted enums into the source table")
 	e2e.EnvWaitFor(s.t, env, 3*time.Minute, "normalize enum", func() bool {
 		return s.checkEnums(srcTableName, dstTableName) == nil
-	})
-
-	env.Cancel()
-
-	e2e.RequireEnvCanceled(s.t, env)
-}
-
-func (s PeerFlowE2ETestSuitePG) Test_Simple_Schema_Changes_PG() {
-	tc := e2e.NewTemporalClient(s.t)
-
-	srcTableName := s.attachSchemaSuffix("test_simple_schema_changes")
-	dstTableName := s.attachSchemaSuffix("test_simple_schema_changes_dst")
-
-	_, err := s.Conn().Exec(context.Background(), fmt.Sprintf(`
-		CREATE TABLE IF NOT EXISTS %s (
-			id BIGINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
-			c1 BIGINT
-		);
-	`, srcTableName))
-	require.NoError(s.t, err)
-
-	connectionGen := e2e.FlowConnectionGenerationConfig{
-		FlowJobName:      s.attachSuffix("test_simple_schema_changes"),
-		TableNameMapping: map[string]string{srcTableName: dstTableName},
-		Destination:      s.peer,
-	}
-
-	flowConnConfig := connectionGen.GenerateFlowConnectionConfigs()
-	flowConnConfig.MaxBatchSize = 1
-
-	// wait for PeerFlowStatusQuery to finish setup
-	// and then insert and mutate schema repeatedly.
-	env := e2e.ExecutePeerflow(tc, peerflow.CDCFlowWorkflow, flowConnConfig, nil)
-	e2e.SetupCDCFlowStatusQuery(s.t, env, connectionGen)
-
-	// insert first row.
-	_, err = s.Conn().Exec(context.Background(), fmt.Sprintf(`
-		INSERT INTO %s(c1) VALUES ($1)`, srcTableName), 1)
-	e2e.EnvNoError(s.t, env, err)
-	s.t.Log("Inserted initial row in the source table")
-
-	s.WaitForSchema(env, "normalizing first row", srcTableName, dstTableName, "id,c1", &protos.TableSchema{
-		TableIdentifier:   dstTableName,
-		PrimaryKeyColumns: []string{"id"},
-		Columns: []*protos.FieldDescription{
-			{
-				Name:         "id",
-				Type:         string(qvalue.QValueKindInt64),
-				TypeModifier: -1,
-			},
-			{
-				Name:         "c1",
-				Type:         string(qvalue.QValueKindInt64),
-				TypeModifier: -1,
-			},
-			{
-				Name:         "_PEERDB_SYNCED_AT",
-				Type:         string(qvalue.QValueKindTimestamp),
-				TypeModifier: -1,
-			},
-		},
-	})
-
-	// alter source table, add column c2 and insert another row.
-	_, err = s.Conn().Exec(context.Background(), fmt.Sprintf(`
-		ALTER TABLE %s ADD COLUMN c2 BIGINT`, srcTableName))
-	e2e.EnvNoError(s.t, env, err)
-	s.t.Log("Altered source table, added column c2")
-	_, err = s.Conn().Exec(context.Background(), fmt.Sprintf(`
-		INSERT INTO %s(c1,c2) VALUES ($1,$2)`, srcTableName), 2, 2)
-	e2e.EnvNoError(s.t, env, err)
-	s.t.Log("Inserted row with added c2 in the source table")
-
-	s.WaitForSchema(env, "normalizing altered row", srcTableName, dstTableName, "id,c1,c2", &protos.TableSchema{
-		TableIdentifier:   dstTableName,
-		PrimaryKeyColumns: []string{"id"},
-		Columns: []*protos.FieldDescription{
-			{
-				Name:         "id",
-				Type:         string(qvalue.QValueKindInt64),
-				TypeModifier: -1,
-			},
-			{
-				Name:         "c1",
-				Type:         string(qvalue.QValueKindInt64),
-				TypeModifier: -1,
-			},
-			{
-				Name:         "_PEERDB_SYNCED_AT",
-				Type:         string(qvalue.QValueKindTimestamp),
-				TypeModifier: -1,
-			},
-			{
-				Name:         "c2",
-				Type:         string(qvalue.QValueKindInt64),
-				TypeModifier: -1,
-			},
-		},
-	})
-
-	// alter source table, add column c3, drop column c2 and insert another row.
-	_, err = s.Conn().Exec(context.Background(), fmt.Sprintf(`
-		ALTER TABLE %s DROP COLUMN c2, ADD COLUMN c3 BIGINT`, srcTableName))
-	e2e.EnvNoError(s.t, env, err)
-	s.t.Log("Altered source table, dropped column c2 and added column c3")
-	_, err = s.Conn().Exec(context.Background(), fmt.Sprintf(`
-		INSERT INTO %s(c1,c3) VALUES ($1,$2)`, srcTableName), 3, 3)
-	e2e.EnvNoError(s.t, env, err)
-	s.t.Log("Inserted row with added c3 in the source table")
-
-	s.WaitForSchema(env, "normalizing dropped column row", srcTableName, dstTableName, "id,c1,c3", &protos.TableSchema{
-		TableIdentifier:   dstTableName,
-		PrimaryKeyColumns: []string{"id"},
-		Columns: []*protos.FieldDescription{
-			{
-				Name:         "id",
-				Type:         string(qvalue.QValueKindInt64),
-				TypeModifier: -1,
-			},
-			{
-				Name:         "c1",
-				Type:         string(qvalue.QValueKindInt64),
-				TypeModifier: -1,
-			},
-			{
-				Name:         "c2",
-				Type:         string(qvalue.QValueKindInt64),
-				TypeModifier: -1,
-			},
-			{
-				Name:         "_PEERDB_SYNCED_AT",
-				Type:         string(qvalue.QValueKindTimestamp),
-				TypeModifier: -1,
-			},
-			{
-				Name:         "c3",
-				Type:         string(qvalue.QValueKindInt64),
-				TypeModifier: -1,
-			},
-		},
-	})
-
-	// alter source table, drop column c3 and insert another row.
-	_, err = s.Conn().Exec(context.Background(), fmt.Sprintf(`
-		ALTER TABLE %s DROP COLUMN c3`, srcTableName))
-	e2e.EnvNoError(s.t, env, err)
-	s.t.Log("Altered source table, dropped column c3")
-	_, err = s.Conn().Exec(context.Background(), fmt.Sprintf(`
-		INSERT INTO %s(c1) VALUES ($1)`, srcTableName), 4)
-	e2e.EnvNoError(s.t, env, err)
-	s.t.Log("Inserted row after dropping all columns in the source table")
-
-	s.WaitForSchema(env, "normalizing 2nd dropped column row", srcTableName, dstTableName, "id,c1", &protos.TableSchema{
-		TableIdentifier:   dstTableName,
-		PrimaryKeyColumns: []string{"id"},
-		Columns: []*protos.FieldDescription{
-			{
-				Name:         "id",
-				Type:         string(qvalue.QValueKindInt64),
-				TypeModifier: -1,
-			},
-			{
-				Name:         "c1",
-				Type:         string(qvalue.QValueKindInt64),
-				TypeModifier: -1,
-			},
-			{
-				Name:         "_PEERDB_SYNCED_AT",
-				Type:         string(qvalue.QValueKindTimestamp),
-				TypeModifier: -1,
-			},
-			{
-				Name:         "c2",
-				Type:         string(qvalue.QValueKindInt64),
-				TypeModifier: -1,
-			},
-			{
-				Name:         "c3",
-				Type:         string(qvalue.QValueKindInt64),
-				TypeModifier: -1,
-			},
-		},
 	})
 
 	env.Cancel()
