+++ flow/e2e/snowflake/peer_flow_sf_test.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"strings"
 	"testing"
 	"time"
 
@@ -16,7 +15,6 @@ import (
 	"github.com/PeerDB-io/peer-flow/e2e"
 	"github.com/PeerDB-io/peer-flow/e2eshared"
 	"github.com/PeerDB-io/peer-flow/generated/protos"
-	"github.com/PeerDB-io/peer-flow/model/qvalue"
 	peerflow "github.com/PeerDB-io/peer-flow/workflows"
 )
 
@@ -510,218 +508,6 @@ func (s PeerFlowE2ETestSuiteSF) Test_Multi_Table_SF() {
 
 		return count1 == 1 && count2 == 1
 	})
-
-	env.Cancel()
-
-	e2e.RequireEnvCanceled(s.t, env)
-}
-
-func (s PeerFlowE2ETestSuiteSF) Test_Simple_Schema_Changes_SF() {
-	tc := e2e.NewTemporalClient(s.t)
-
-	srcTableName := s.attachSchemaSuffix("test_simple_schema_changes")
-	dstTableName := fmt.Sprintf("%s.%s", s.sfHelper.testSchemaName, "test_simple_schema_changes")
-
-	_, err := s.Conn().Exec(context.Background(), fmt.Sprintf(`
-		CREATE TABLE IF NOT EXISTS %s (
-			id BIGINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
-			c1 BIGINT
-		);
-	`, srcTableName))
-	require.NoError(s.t, err)
-
-	connectionGen := e2e.FlowConnectionGenerationConfig{
-		FlowJobName:      s.attachSuffix("test_simple_schema_changes"),
-		TableNameMapping: map[string]string{srcTableName: dstTableName},
-		Destination:      s.sfHelper.Peer,
-	}
-
-	flowConnConfig := connectionGen.GenerateFlowConnectionConfigs()
-	flowConnConfig.MaxBatchSize = 100
-
-	// wait for PeerFlowStatusQuery to finish setup
-	// and then insert and mutate schema repeatedly.
-	env := e2e.ExecutePeerflow(tc, peerflow.CDCFlowWorkflow, flowConnConfig, nil)
-	e2e.SetupCDCFlowStatusQuery(s.t, env, connectionGen)
-	_, err = s.Conn().Exec(context.Background(), fmt.Sprintf(`
-		INSERT INTO %s(c1) VALUES ($1)`, srcTableName), 1)
-	e2e.EnvNoError(s.t, env, err)
-	s.t.Log("Inserted initial row in the source table")
-
-	e2e.EnvWaitForEqualTables(env, s, "normalize reinsert", "test_simple_schema_changes", "id,c1")
-
-	expectedTableSchema := &protos.TableSchema{
-		TableIdentifier: strings.ToUpper(dstTableName),
-		Columns: []*protos.FieldDescription{
-			{
-				Name:         "ID",
-				Type:         string(qvalue.QValueKindNumeric),
-				TypeModifier: -1,
-			},
-			{
-				Name:         "C1",
-				Type:         string(qvalue.QValueKindNumeric),
-				TypeModifier: -1,
-			},
-			{
-				Name:         "_PEERDB_IS_DELETED",
-				Type:         string(qvalue.QValueKindBoolean),
-				TypeModifier: -1,
-			},
-			{
-				Name:         "_PEERDB_SYNCED_AT",
-				Type:         string(qvalue.QValueKindTimestamp),
-				TypeModifier: -1,
-			},
-		},
-	}
-	output, err := s.connector.GetTableSchema(context.Background(), &protos.GetTableSchemaBatchInput{
-		TableIdentifiers: []string{dstTableName},
-	})
-	e2e.EnvNoError(s.t, env, err)
-	e2e.EnvTrue(s.t, env, e2e.CompareTableSchemas(expectedTableSchema, output.TableNameSchemaMapping[dstTableName]))
-
-	// alter source table, add column c2 and insert another row.
-	_, err = s.Conn().Exec(context.Background(), fmt.Sprintf(`
-		ALTER TABLE %s ADD COLUMN c2 BIGINT`, srcTableName))
-	e2e.EnvNoError(s.t, env, err)
-	s.t.Log("Altered source table, added column c2")
-	_, err = s.Conn().Exec(context.Background(), fmt.Sprintf(`
-		INSERT INTO %s(c1,c2) VALUES ($1,$2)`, srcTableName), 2, 2)
-	e2e.EnvNoError(s.t, env, err)
-	s.t.Log("Inserted row with added c2 in the source table")
-
-	// verify we got our two rows, if schema did not match up it will error.
-	e2e.EnvWaitForEqualTables(env, s, "normalize altered row", "test_simple_schema_changes", "id,c1,c2")
-	expectedTableSchema = &protos.TableSchema{
-		TableIdentifier: strings.ToUpper(dstTableName),
-		Columns: []*protos.FieldDescription{
-			{
-				Name:         "ID",
-				Type:         string(qvalue.QValueKindNumeric),
-				TypeModifier: -1,
-			},
-			{
-				Name:         "C1",
-				Type:         string(qvalue.QValueKindNumeric),
-				TypeModifier: -1,
-			},
-			{
-				Name:         "_PEERDB_SYNCED_AT",
-				Type:         string(qvalue.QValueKindTimestamp),
-				TypeModifier: -1,
-			},
-			{
-				Name:         "C2",
-				Type:         string(qvalue.QValueKindNumeric),
-				TypeModifier: -1,
-			},
-		},
-	}
-	output, err = s.connector.GetTableSchema(context.Background(), &protos.GetTableSchemaBatchInput{
-		TableIdentifiers: []string{dstTableName},
-	})
-	e2e.EnvNoError(s.t, env, err)
-	e2e.EnvTrue(s.t, env, e2e.CompareTableSchemas(expectedTableSchema, output.TableNameSchemaMapping[dstTableName]))
-	e2e.EnvEqualTables(env, s, "test_simple_schema_changes", "id,c1,c2")
-
-	// alter source table, add column c3, drop column c2 and insert another row.
-	_, err = s.Conn().Exec(context.Background(), fmt.Sprintf(`
-		ALTER TABLE %s DROP COLUMN c2, ADD COLUMN c3 BIGINT`, srcTableName))
-	e2e.EnvNoError(s.t, env, err)
-	s.t.Log("Altered source table, dropped column c2 and added column c3")
-	_, err = s.Conn().Exec(context.Background(), fmt.Sprintf(`
-		INSERT INTO %s(c1,c3) VALUES ($1,$2)`, srcTableName), 3, 3)
-	e2e.EnvNoError(s.t, env, err)
-	s.t.Log("Inserted row with added c3 in the source table")
-
-	// verify we got our two rows, if schema did not match up it will error.
-	e2e.EnvWaitForEqualTables(env, s, "normalize dropped c2 column", "test_simple_schema_changes", "id,c1,c3")
-	expectedTableSchema = &protos.TableSchema{
-		TableIdentifier: strings.ToUpper(dstTableName),
-		Columns: []*protos.FieldDescription{
-			{
-				Name:         "ID",
-				Type:         string(qvalue.QValueKindNumeric),
-				TypeModifier: -1,
-			},
-			{
-				Name:         "C1",
-				Type:         string(qvalue.QValueKindNumeric),
-				TypeModifier: -1,
-			},
-			{
-				Name:         "_PEERDB_SYNCED_AT",
-				Type:         string(qvalue.QValueKindTimestamp),
-				TypeModifier: -1,
-			},
-			{
-				Name:         "C2",
-				Type:         string(qvalue.QValueKindNumeric),
-				TypeModifier: -1,
-			},
-			{
-				Name:         "C3",
-				Type:         string(qvalue.QValueKindNumeric),
-				TypeModifier: -1,
-			},
-		},
-	}
-	output, err = s.connector.GetTableSchema(context.Background(), &protos.GetTableSchemaBatchInput{
-		TableIdentifiers: []string{dstTableName},
-	})
-	e2e.EnvNoError(s.t, env, err)
-	e2e.EnvTrue(s.t, env, e2e.CompareTableSchemas(expectedTableSchema, output.TableNameSchemaMapping[dstTableName]))
-	e2e.EnvEqualTables(env, s, "test_simple_schema_changes", "id,c1,c3")
-
-	// alter source table, drop column c3 and insert another row.
-	_, err = s.Conn().Exec(context.Background(), fmt.Sprintf(`
-		ALTER TABLE %s DROP COLUMN c3`, srcTableName))
-	e2e.EnvNoError(s.t, env, err)
-	s.t.Log("Altered source table, dropped column c3")
-	_, err = s.Conn().Exec(context.Background(), fmt.Sprintf(`
-		INSERT INTO %s(c1) VALUES ($1)`, srcTableName), 4)
-	e2e.EnvNoError(s.t, env, err)
-	s.t.Log("Inserted row after dropping all columns in the source table")
-
-	// verify we got our two rows, if schema did not match up it will error.
-	e2e.EnvWaitForEqualTables(env, s, "normalize dropped c3 column", "test_simple_schema_changes", "id,c1")
-	expectedTableSchema = &protos.TableSchema{
-		TableIdentifier: strings.ToUpper(dstTableName),
-		Columns: []*protos.FieldDescription{
-			{
-				Name:         "ID",
-				Type:         string(qvalue.QValueKindNumeric),
-				TypeModifier: -1,
-			},
-			{
-				Name:         "C1",
-				Type:         string(qvalue.QValueKindNumeric),
-				TypeModifier: -1,
-			},
-			{
-				Name:         "_PEERDB_SYNCED_AT",
-				Type:         string(qvalue.QValueKindTimestamp),
-				TypeModifier: -1,
-			},
-			{
-				Name:         "C2",
-				Type:         string(qvalue.QValueKindNumeric),
-				TypeModifier: -1,
-			},
-			{
-				Name:         "C3",
-				Type:         string(qvalue.QValueKindNumeric),
-				TypeModifier: -1,
-			},
-		},
-	}
-	output, err = s.connector.GetTableSchema(context.Background(), &protos.GetTableSchemaBatchInput{
-		TableIdentifiers: []string{dstTableName},
-	})
-	e2e.EnvNoError(s.t, env, err)
-	e2e.EnvTrue(s.t, env, e2e.CompareTableSchemas(expectedTableSchema, output.TableNameSchemaMapping[dstTableName]))
-	e2e.EnvEqualTables(env, s, "test_simple_schema_changes", "id,c1")
 
 	env.Cancel()
