+++ flow/e2e/test_utils.go
@@ -22,6 +22,7 @@ import (
 	"go.temporal.io/sdk/converter"
 	"go.temporal.io/sdk/temporal"
 
+	"github.com/PeerDB-io/peer-flow/connectors"
 	connpostgres "github.com/PeerDB-io/peer-flow/connectors/postgres"
 	connsnowflake "github.com/PeerDB-io/peer-flow/connectors/snowflake"
 	"github.com/PeerDB-io/peer-flow/connectors/utils"
@@ -55,6 +56,7 @@ type RowSource interface {
 type GenericSuite interface {
 	RowSource
 	Peer() *protos.Peer
+	DestinationConnector() connectors.Connector
 	DestinationTable(table string) string
 }
 
@@ -112,13 +114,17 @@ func RequireEqualTables(suite RowSource, table string, cols string) {
 }
 
 func EnvEqualTables(env WorkflowRun, suite RowSource, table string, cols string) {
+	EnvEqualTablesWithNames(env, suite, table, table, cols)
+}
+
+func EnvEqualTablesWithNames(env WorkflowRun, suite RowSource, srcTable string, dstTable string, cols string) {
 	t := suite.T()
 	t.Helper()
 
-	pgRows, err := GetPgRows(suite.Connector(), suite.Suffix(), table, cols)
+	pgRows, err := GetPgRows(suite.Connector(), suite.Suffix(), srcTable, cols)
 	EnvNoError(t, env, err)
 
-	rows, err := suite.GetRows(table, cols)
+	rows, err := suite.GetRows(dstTable, cols)
 	EnvNoError(t, env, err)
 
 	EnvEqualRecordBatches(t, env, pgRows, rows)
@@ -517,6 +523,19 @@ func GetOwnersSelectorStringsSF() [2]string {
 		}
 	}
 	return [2]string{strings.Join(pgFields, ","), strings.Join(sfFields, ",")}
+}
+
+func ExpectedDestinationIdentifier(s GenericSuite, ident string) string {
+	switch s.DestinationConnector().(type) {
+	case *connsnowflake.SnowflakeConnector:
+		return strings.ToUpper(ident)
+	default:
+		return ident
+	}
+}
+
+func ExpectedDestinationTableName(s GenericSuite, table string) string {
+	return ExpectedDestinationIdentifier(s, s.DestinationTable(table))
 }
 
 type testWriter struct {
